pyproject.toml
[tool.poetry]
name = "chatter-base"
version = "0.1.0"
description = ""
authors = ["NekrodNIK"]
license = "MIT"
readme = "README.md"
packages = [{include = "chatter_base"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
python-jose = "^3.3.0"
sqlalchemy = "^2.0.8"
uvicorn = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
